Update library entity types
