+++ unified-codes/libs/data/src/lib/types/EntitySearch.ts
@@ -1,0 +1,12 @@
+import { EEntityType } from './Entity';
+
+import IEntitySort from './EntitySort';
+
+export interface IEntitySearch {
+    code: string;
+    description: string;
+    orderBy: IEntitySort;
+    type: EEntityType | EEntityType[];
+  }
+
+  export default IEntitySearch;
