+++ unified-codes/libs/data/src/lib/types/EntitySort.ts
@@ -1,0 +1,8 @@
+import { EEntityField } from './Entity';
+
+export interface IEntitySort {
+    field: EEntityField;
+    descending: boolean;
+  }
+
+export default IEntitySort;
