+++ unified-codes/libs/data/src/lib/types/index.ts
@@ -9,3 +9,5 @@ export * from './IdentityProvider';
 export * from './JWT';
 export * from './Property';
 export * from './User';
+export * from './EntitySearch';
+export * from './EntitySort';
